Implement splitAtDelimiters based on KaTeX
The `Latexifier` is not yet complete, however, the bulk of the code is
(i.e. `splitAtDelimiters` and `findEndOfMath`).

The code is a fairly naive rewrite from JS to Rust so I imagine there
are quite a few points that could be drastically improved upon. I will
add comments and documentation once I fully understand the code I'm
porting over.