Fix the `hyde new` subcommand
Rather embarrassingly, the `hyde new` command did not work because
`hyde.toml` had not been created in that directory yet, and a newly
created Hyde would also not build successfully after initialisation
because the `posts/` directory was not created. These issues have now
been fixed such that `hyde new` now creates the `hyde.toml` file if it
does not exist, and also creates a new, empty `posts/` sub-directory in
the project.